Move GmsReader code to parser.py
